[#4908] Python 2.7.15 update. (#104)
* Updated Python sources to 2.7.15.

* Updated Python Windows installer to 2.7.15.

* Updated deps for Ubuntu 18.04.

* Updated chevah_build and src/python/ patches for new Python version.

* A second AIX-specific hack for ld_so_aix woes.

* Changes after own review.

* Show exit code if not 0, when $DEBUG is set.

* Soft-fail on chevah-keycert tests for now.

* Removed chevah-keycert tests through reseting from a past version.

* Do not try compat tests on unsupported platforms.

* Minor Solaris 10 improvements.

+++ functions.sh
@@ -126,6 +126,9 @@ execute() {
     #Make sure $@ is called in quotes as otherwise it will not work.
     "$@"
     exit_code=$?
+    if [ $DEBUG -ne 0 ]; then
+        echo "Exit code was: $exit_code"
+    fi
     if [ $exit_code -ne 0 ]; then
         echo "PWD :" `pwd`
         echo "Fail:" $@
@@ -290,6 +293,10 @@ aix_ld_hack() {
     custom_python_dir="/lib/$PYTHON_VERSION"
     case $* in
         init)
+            if [ -e "$custom_python_dir" ]; then
+                (>&2 echo "$custom_python_dir already exists, not proceeding!")
+                exit 99
+            fi
             execute sudo mkdir -p "$custom_python_dir"/config
             execute sudo cp "$INSTALL_FOLDER"/lib/"$PYTHON_VERSION"/config/* \
                 "$custom_python_dir"/config/
